fix: don't record failures on context cancellation
Gate the case where the retriever (parent) cancels a protocol retriever because
another has completed successfully. If a retrieval fails in this case, it's not
a failure that we should either report in the event system, or record against
the SP's performance stats.

Fixes: https://github.com/filecoin-project/lassie/issues/364

I believe this fixes #364, but it's hard to reproduce reliably so I can't be
sure.