fix(cli): clean up path before passing it on
dba04a25 cleaned up pathing elsewhere, removing "/" prefixes, but left this undone. It means that a "/" prefix passes through the system and we end up with "//"; which is fine, except for the HTTP retriever which will fetch with a "//" prefix and be corrected by the server's `Location` header. This change does a parse, just like the daemon handler does, and strips out a "/" prefixe if it exists.