Export IpfsHandler
Saturn / other clients want to experiment with other http servers making using of lassie.

The IPFS Handler in the server/http package contains application logic for how to parse / render URLs that is reusable.
The specific semantics of the HTTP server that this handler is currently attached to are specific to the lassie daemon mode - e.g. it may not be wnated to have the pprof debug endpoints on the same mux, or extra
middleware like caching may be desired.